Fix bug: failure if ship has left any graphs
The metadata-store maintains graph associations
for graphs a ship has joined and then left.
However, these graphs no longer exist.

If you attempt to scry a graph that does not
exist, the attempt fails and the caller crashes.

To remedy, check each graph-association-sourced
resource against a set of resources scryed from
graph-store, which is up-to-date.